+++ src/g_hexen/a_hexenspecialdecs.cpp
@@ -108,8 +108,8 @@ DEFINE_ACTION_FUNCTION(AActor, A_PotteryCheck)
 		if (playeringame[i])
 		{
 			AActor *pmo = players[i].mo;
-			if (P_CheckSight (self, pmo) && (abs (R_PointToAngle2 (pmo->x,
-				pmo->y, self->x, self->y) - pmo->angle) <= ANGLE_45))
+			if (P_CheckSight (self, pmo) && (abs (int32(R_PointToAngle2 (pmo->x,
+				pmo->y, self->x, self->y) - pmo->angle)) <= ANGLE_45))
 			{ // Previous state (pottery bit waiting state)
 				self->SetState (self->state - 1);
 				return;
